+++ client/js/log-in.js
@@ -1,8 +1,20 @@
 const logInForm = document.getElementById('log-in-form')
 
+const processAuth = (res) => {
+
+	const username = res.username
+	const admin = res.admin
+
+	if (username !== undefined) {
+		window.location = `/homepage.html?admin=${admin}`
+	}
+	else console.log(res)
+}
+
 logInForm.addEventListener('submit', e => {
 	e.preventDefault()
-
+	console.log("login")
+	
 	const username = document.getElementById('username')
 	const password = document.getElementById('password')
 
@@ -11,16 +23,12 @@ logInForm.addEventListener('submit', e => {
 		password: password.value,
 	}
 
-	fetch('http://localhost:3030/log-in', {
+	fetch('http://localhost:3030/route-session/log-in', {
 		method: 'post',
-		headers: {
-			'Content-Type': 'application/json',
-		},
+		headers: { 'Content-Type': 'application/json' },
 		'credentials': 'include',
 		body: JSON.stringify(form)
 	})
-	.then(res => {
-		res.json() 
-		window.location = '/homepage.html'
-	})
+	.then(res => res.json())
+	.then(processAuth)
 })
