--- client/js/tools/tool-box-rss.js
@@ -1,41 +1,0 @@
-// Change XML en JSON -- source : https://davidwalsh.name/convert-xml-json
-
-const xmlToJson = (xml) => {
-	
-	// Create the return object
-	var obj = {}
-
-	if (xml.nodeType == 1) { // element
-		// do attributes
-		if (xml.attributes.length > 0) {
-		obj["@attributes"] = {}
-			for (let j = 0 ; j < xml.attributes.length ; j++) {
-				var attribute = xml.attributes.item(j)
-				obj["@attributes"][attribute.nodeName] = attribute.nodeValue
-			}
-		}
-	} else if (xml.nodeType == 3) { // text
-		obj = xml.nodeValue
-	}
-
-	// do children
-	if (xml.hasChildNodes()) {
-
-		for(let i = 0 ; i < xml.childNodes.length ; i++) {
-			var item = xml.childNodes.item(i)
-			var nodeName = item.nodeName
-			if (typeof(obj[nodeName]) == "undefined") {
-				obj[nodeName] = xmlToJson(item)
-			} else {
-				if (typeof(obj[nodeName].push) == "undefined") {
-					var old = obj[nodeName]
-					obj[nodeName] = []
-					obj[nodeName].push(old)
-				}
-				obj[nodeName].push(xmlToJson(item))
-			}
-        }
-        
-	}
-	return obj
-}
